Phase A9: split reconcile.rs into slot.rs and diff.rs
reconcile.rs at 534 lines crossed the 500-line soft-cap. Natural seam:
GenerationSupervisor (supervisor logic) stays in reconcile.rs; slot
lifecycle management moves to reconcile/slot.rs; snapshot diffing moves
to reconcile/diff.rs. No behavior changes.
